add harm categories
